💄 Design : 학생 관리 페이지 CSS 수정
- 학생 관리 페이지 CSS 수정
- 각 행에서 포인트 아이콘 삭제
- 하단에 포인트 지급 버튼 추가
