+++ src/components/manageLearner/LearnerItem.tsx
@@ -1,9 +1,7 @@
 import { ILearnerItem } from "@/interface/learnerItem";
-import { paymentPointModalState } from "@/states/confirmModalState";
 import { selectedLearner } from "@/states/manageLearner";
-import PointIcon from "@assets/svg/pointIcon.svg?react";
 import { useEffect } from "react";
-import { useRecoilState, useSetRecoilState } from "recoil";
+import { useRecoilState } from "recoil";
 
 interface ILearnerInfoProps {
 	readonly learnerInfo: ILearnerItem;
@@ -12,7 +10,6 @@ interface ILearnerInfoProps {
 const LearnerItem = ({ learnerInfo }: ILearnerInfoProps) => {
 	const [selectedLearners, setSelectedLearners] =
 		useRecoilState(selectedLearner);
-	const setIsOpenPaymentPoint = useSetRecoilState(paymentPointModalState);
 
 	const onClickLearnerHandler = () => {
 		const index = selectedLearners.findIndex((learnerId) => {
@@ -26,10 +23,6 @@ const LearnerItem = ({ learnerInfo }: ILearnerInfoProps) => {
 			deepCopySelectedLearners.splice(index, 1);
 			setSelectedLearners(deepCopySelectedLearners);
 		}
-	};
-
-	const openPaymentPointModalHandler = () => {
-		setIsOpenPaymentPoint(true);
 	};
 
 	useEffect(() => {
@@ -63,13 +56,6 @@ const LearnerItem = ({ learnerInfo }: ILearnerInfoProps) => {
 						<span> ({learnerInfo.learnerId}) </span>
 					</p>
 				</div>
-				<button
-					type="button"
-					className="mx-4 flex h-6 w-6 items-center justify-center rounded-full bg-tangerine"
-					onClick={openPaymentPointModalHandler}
-				>
-					<PointIcon width={9} height={12} />
-				</button>
 			</div>
 		</div>
 	);
