+++ src/components/manageLearner/ManageLearnerContainer.tsx
@@ -3,7 +3,7 @@ import {
 	paymentPointModalState,
 } from "@/states/confirmModalState";
 import { useRecoilValue } from "recoil";
-import DeleteLearnerButton from "./DeleteLearnerButton";
+import DeleteLearnerButton from "./DeleteLearnerAndPointButton";
 import DeleteLearnerModal from "./modal/DeleteLearnerModal";
 import LearnerList from "./LearnerList";
 import LearnerSearch from "./LearnerSearch";
